Client: encapsulation
