+++ .idea/compiler.xml
@@ -8,7 +8,6 @@
         <outputRelativeToContentRoot value="true" />
         <module name="remote-storage-server" />
         <module name="remote-storage-client" />
-        <module name="remote-storage-common" />
       </profile>
     </annotationProcessing>
   </component>
