+++ remote-storage-client/pom.xml
@@ -37,12 +37,6 @@
             <version>2.13.3</version>
         </dependency>
         <dependency>
-            <groupId>kriuchkov.maksim</groupId>
-            <artifactId>remote-storage-common</artifactId>
-            <version>1.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
             <groupId>org.openjfx</groupId>
             <artifactId>javafx-controls</artifactId>
             <version>14</version>
@@ -51,6 +45,12 @@
             <groupId>org.openjfx</groupId>
             <artifactId>javafx-fxml</artifactId>
             <version>14</version>
+        </dependency>
+        <dependency>
+            <groupId>kriuchkov.maksim</groupId>
+            <artifactId>remote-storage-common</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
         </dependency>
     </dependencies>
