--- remote-storage-client/src/main/java/kriuchkov/maksim/client/ClientFileService.java
@@ -1,16 +1,0 @@
-package kriuchkov.maksim.client;
-
-import kriuchkov.maksim.common.FileService;
-
-public class ClientFileService extends FileService {
-
-    private static final ClientFileService instance = new ClientFileService();
-
-    public static ClientFileService getInstance() {
-        return instance;
-    }
-
-    public void doStore() throws Exception {
-        sendFile(dataSource, NetworkHandler.getInstance().getChannel(), null);
-    }
-}
