+++ remote-storage-common/src/main/java/kriuchkov/maksim/common/Protocol.java
@@ -5,5 +5,4 @@ public class Protocol {
     public static final byte DATA_SIGNAL_BYTE = 2;
 
     public static final int MAX_FRAME_BODY_LENGTH = 0x20000;
-
 }
